+++ sdl2/ext/surface.py
@@ -68,16 +68,50 @@ def _create_surface(size, fill=None, fmt="ARGB8888", errname="SDL"):
 
 
 def subsurface(surface, area):
-    """Creates a surface from a part of another surface.
+    """Creates a new :obj:`~sdl2.SDL_Surface` from a part of another surface.
 
-    The two surfaces share pixel data. The subsurface *must not* be used after
-    its parent has been freed!
+    Surfaces created with this function will share pixel data with the original
+    surface, meaning that any modifications to one surface will result in
+    modifications to the other.
+
+    .. warning::
+       Because subsurfaces share pixel data with their parent surface, they
+       *cannot* be used after the parent surface is freed. Doing so will
+       almost certainly result in a segfault.
+
+    Args:
+        surface (:obj:`~sdl2.SDL_Surface`): The parent surface from which
+            new sub-surface should be created.
+        area (:obj:`SDL_Rect`, tuple): The ``(x, y, w, h)`` subset of the parent
+            surface to use for the new surface, where ``x, y`` are the pixel
+            coordinates of the top-left corner of the rectangle and ``w, h`` are
+            its width and height (in pixels). Can also be specified as an
+            :obj:`SDL_Rect`.
+
+    Returns:
+        :obj:`~sdl2.SDL_Surface`: The newly-created subsurface.
+
     """
-    surface_format = surface.format[0]
-    subpixels = (surface.pixels + surface.pitch*area[1] +
-                 surface_format.BytesPerPixel*area[0])
-    return surf.SDL_CreateRGBSurfaceFrom(subpixels, area[2], area[3],
-                                    surface_format.BitsPerPixel,
-                                    surface.pitch, surface_format.Rmask,
-                                    surface_format.Gmask, surface_format.Bmask,
-                                    surface_format.Amask)[0]
+    if not isinstance(surface, surf.SDL_Surface):
+        if "SDL_Surface" in str(type(surface)):
+            surface = surface.contents
+        else:
+            e = "'surface' must be an SDL_Surface (got {0})"
+            raise TypeError(e.format(type(surface)))
+
+    x, y, w, h = _get_rect_tuple(area, argname="area")
+    if x + w > surface.w or y + h > surface.h:
+        e = "The specified area {0} exceeds the bounds of the parent surface "
+        e += str((surface.w, surface.h))
+        raise ValueError(e.format(str(area)))
+
+    fmt = surface.format[0]
+    bpp = fmt.BitsPerPixel
+    subpixels = (surface.pixels + surface.pitch * y + fmt.BytesPerPixel * x)
+    subsurf = surf.SDL_CreateRGBSurfaceFrom(
+        subpixels, w, h, bpp, surface.pitch, fmt.Rmask, fmt.Gmask, fmt.Bmask, fmt.Amask
+    )
+    if not subsurf:
+        raise_sdl_err("creating the subsurface")
+        
+    return subsurf.contents